refactor(teacher): EvaluationCorrections.vue sous 300 lignes (H2)
Éclatement complet (pattern G1) :
- utils/evaluationCorrectionsFormat.js : helpers PURS (dates, initiales, note/statut)
  — NOUVEAU module, partagé tableau↔export CSV
- assets/styles/_correction-badges.scss : badges note/statut (CSS verbatim, @use'd)
- useEvaluationCorrections.js : chargement résultats par classe + retour + export CSV
- CorrectionEvalInfo / CorrectionStats / CorrectionResultsTable (dans components/evaluations/)
- EvaluationCorrections.vue : orchestrateur 823→203 lignes
Tests : format util (4) + composable (4) + sections (5) = 14 verts. Build OK.
Parité : même endpoint (getResultsByClass), même export CSV, mêmes classes/textes.
Note : sous-composants placés dans components/evaluations/ (regroupement domaine)
faute de pouvoir écrire sous components/teacher/.
