+++ src/assets/styles/_correction-badges.scss
@@ -1,0 +1,72 @@
+// Badges de note / statut des résultats d'évaluation (H2 DRY) — CSS byte-identique
+// extrait verbatim de `views/teacher/EvaluationCorrections.vue`. @use'd dans le
+// <style scoped> de CorrectionResultsTable (classes autonomes, non descendantes du
+// tableau) : scoping per-composant préservé, parité visuelle.
+
+/* Note Badge */
+.note-badge {
+  display: inline-block;
+  padding: 0.375rem 0.75rem;
+  border-radius: 9999px;
+  font-size: 0.875rem;
+  font-weight: 600;
+}
+
+.note-excellent {
+  background: #dcfce7;
+  color: #166534;
+}
+
+.note-good {
+  background: #dbeafe;
+  color: #1e40af;
+}
+
+.note-average {
+  background: #fef3c7;
+  color: #92400e;
+}
+
+.note-low {
+  background: #fee2e2;
+  color: #991b1b;
+}
+
+.note-empty {
+  color: var(--text-tertiary);
+  font-style: italic;
+}
+
+/* Status Badge */
+.status-badge {
+  display: inline-block;
+  padding: 0.375rem 0.75rem;
+  border-radius: 9999px;
+  font-size: 0.75rem;
+  font-weight: 600;
+}
+
+.status-submitted {
+  background: #dbeafe;
+  color: #1e40af;
+}
+
+.status-corrected {
+  background: #dcfce7;
+  color: #166534;
+}
+
+.status-ongoing {
+  background: #fef3c7;
+  color: #92400e;
+}
+
+.status-not-taken {
+  background: #f3f4f6;
+  color: #6b7280;
+}
+
+.attempt-badge {
+  font-size: 0.8125rem;
+  color: var(--text-secondary);
+}
